extracting TreeNode
