+++ Library/DataStructure/TreeNode.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+
+namespace Library.DataStructure
+{
+    internal class TreeNode<T>
+    {
+        internal T Data;
+        internal TreeNode<T> Father;
+        internal List<TreeNode<T>> Next = new List<TreeNode<T>>();
+        public TreeNode(T data)
+        {
+            Data = data;
+        }
+    }
+}
